# perf(core): skip unnecessary exists checks