**What's the problem this PR addresses?**

When setting up `Workspace` and `Manifest` instances Yarn does an unnecessary "file exists" check.

**How did you fix it?**

Remove them.

**Benchmark results**

```sh
# Berry repo (Ran in WSL but stored on the Windows side (/mnt/c/))

YARN_IGNORE_PATH=1 hyperfine -w 1\
                    "node ./master.cjs exec pwd"\
                    "node ./skip-exists-sync.cjs exec pwd"\
                    "node ./skip-exists-sync-and-promise.cjs exec pwd"
Benchmark 1: node ./master.cjs exec pwd
  Time (mean ± σ):     945.1 ms ±  10.3 ms    [User: 786.6 ms, System: 112.4 ms]
  Range (min … max):   932.9 ms … 968.9 ms    10 runs

Benchmark 2: node ./skip-exists-sync.cjs exec pwd
  Time (mean ± σ):     897.0 ms ±   5.8 ms    [User: 781.2 ms, System: 106.7 ms]
  Range (min … max):   888.0 ms … 906.4 ms    10 runs

Benchmark 3: node ./skip-exists-sync-and-promise.cjs exec pwd
  Time (mean ± σ):     846.3 ms ±   6.7 ms    [User: 770.4 ms, System: 103.8 ms]
  Range (min … max):   831.7 ms … 854.7 ms    10 runs

Summary
  'node ./skip-exists-sync-and-promise.cjs exec pwd' ran
    1.06 ± 0.01 times faster than 'node ./skip-exists-sync.cjs exec pwd'
    1.12 ± 0.01 times faster than 'node ./master.cjs exec pwd'

# Next.js repo with all examples included (WSL)

YARN_IGNORE_PATH=1 hyperfine -w 1\
                    "node ./master.cjs exec pwd"\
                    "node ./skip-exists-sync.cjs exec pwd"\
                    "node ./skip-exists-sync-and-promise.cjs exec pwd"
Benchmark 1: node ./master.cjs exec pwd
  Time (mean ± σ):      1.417 s ±  0.048 s    [User: 1.627 s, System: 0.232 s]
  Range (min … max):    1.386 s …  1.548 s    10 runs

Benchmark 2: node ./skip-exists-sync.cjs exec pwd
  Time (mean ± σ):      1.395 s ±  0.010 s    [User: 1.604 s, System: 0.229 s]
  Range (min … max):    1.372 s …  1.407 s    10 runs

Benchmark 3: node ./skip-exists-sync-and-promise.cjs exec pwd
  Time (mean ± σ):      1.382 s ±  0.007 s    [User: 1.588 s, System: 0.221 s]
  Range (min … max):    1.372 s …  1.392 s    10 runs

Summary
  'node ./skip-exists-sync-and-promise.cjs exec pwd' ran
    1.01 ± 0.01 times faster than 'node ./skip-exists-sync.cjs exec pwd'
    1.03 ± 0.03 times faster than 'node ./master.cjs exec pwd'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.